Sessions

+++ views/film-single.tpl
@@ -14,8 +14,11 @@
 			<div class="card__header">
 				<h4 class="title-4"><?=$film['title']?></h4>
 				<div>
+				<?php if ( isset($_SESSION['user']) ) {
+					if ( $_SESSION['user'] == 'admin' ) { ?>
 					<a class="button button--editsmall mb-20" href="edit.php?action=edit&id=<?=$film['id']?>">Редактировать</a>
 					<a class="button button--removesmall" href="index.php?action=delete&id=<?=$film['id']?>">Удалить</a>
+				<?php }	} ?>
 				</div>
 			</div>
 			<div class="badge"><?=$film['genre']?></div>
